feat: wrap comint with ink-play derived mode
This wraps the comint-mode buffer used for running ink-play with a
derived mode called ink-play-mode, which adds a single keybinding for
jumping to the docs.

This isn't really necessary, but while reading the manual and testing
out examples, I was running examples and then wanting to jump back to
the docs. It's certainly possible to just set a keybinding elsewhere,
jump via buffer, or use `M-x ink-display-manual`, but I've seen comint
wrapped a few different ways recently and had never done it myself, so
took a shot at it here.

Ink-play mode could be expanded to rewind choices or restart the last
knot/stitch played at some point. Another idea could be writing a
binding to jump to the current choice/prompt from the 'play' buffer.

Just some thoughts, feel free to take it or leave it. Many thanks for
the work on this mode, it has made learning ink completely within emacs
a great experience.