feat: supports 'replay' of last file and knot from ink-play
Adds ink-play and ink-play-knot bindings to ink-play-mode, allowing for
quick retry/replay behavior without switching buffers.

This expands the ink-play function to support being run from both an
ink-mode and ink-play-mode context. Each run updates a local var with
the used file (and knot, if indicated). Subsequent runs from an ink-mode
file will update the cache vars, while runs from ink-play-mode will use
them to re-execute the same path.

Adds comments to track the logic in ink-play, and expands on the docs
for introduced vars somewhat.

As implemented, there is a chance a user calls ink-play-knot, then
switches files, calls ink-play, and then attempts to re-play
ink-play-knot directly, and we end up attempting a knot that does not
exist in the new file... This feels like an edge case we could
handle, but it's probably rare/accidental, so I've left it as is -
doesn't quite seem worth the complexity. Maybe there's some simple check
that could be made? Leaving it for now.